Moved GUI-Unspecific stuff from CxxDocument
Separated out all non-GUI-Relevant functions from
nc::gui::CxxDocument into nc::core::CxxDocument. This
allows using the CxxDocument range features without
the GUI.

nc::gui::CxxDocument now inherits from QTextDocument
as well as from nc::core::CxxDocument.
