+++ src/nc/gui/CxxDocument.h
@@ -36,6 +36,7 @@
 #include <nc/common/Types.h>
 
 #include <nc/core/RangeTree.h>
+#include <nc/core/CxxDocument.h>
 
 namespace nc {
 
@@ -66,7 +67,7 @@ namespace gui {
 /**
  * Text document containing C++ listing.
  */
-class CxxDocument: public QTextDocument {
+class CxxDocument: public QTextDocument, public core::CxxDocument {
     Q_OBJECT
 
     std::shared_ptr<const core::Context> context_;
@@ -87,54 +88,6 @@ public:
     CxxDocument(QObject *parent = nullptr, std::shared_ptr<const core::Context> context = nullptr);
 
     /**
-     * \return Pointer to the deepest tree node at the given position. Can be nullptr.
-     */
-    const core::likec::TreeNode *getLeafAt(int position) const;
-
-    /**
-     * \return List of valid pointers to the nodes fully contained in the given range.
-     */
-    std::vector<const core::likec::TreeNode *> getNodesIn(const Range<int> &range) const;
-
-    /**
-     * \param node Valid pointer to a tree node.
-     *
-     * \return Text range occupied by this node.
-     */
-    Range<int> getRange(const core::likec::TreeNode *node) const;
-
-    /**
-     * \param instruction Valid pointer to an instruction.
-     * \param[out] result List of ranges occupied by the nodes generated from this instruction.
-     */
-    void getRanges(const core::arch::Instruction *instruction, std::vector<Range<int>> &result) const;
-
-    /**
-     * \param declaration Valid pointer to a declaration tree node.
-     *
-     * \return All the tree nodes using this declaration.
-     */
-    const std::vector<const core::likec::TreeNode *> &getUses(const core::likec::Declaration *declaration) const {
-        assert(declaration != nullptr);
-        return nc::find(declaration2uses_, declaration);
-    }
-
-    /**
-     * \param declaration Valid pointer to a label declaration node.
-     *
-     * \return Pointer to the matching label statement. Can be nullptr.
-     */
-    const core::likec::LabelStatement *getLabelStatement(const core::likec::LabelDeclaration *declaration) const {
-        assert(declaration != nullptr);
-        return nc::find(label2statement_, declaration);
-    }
-
-    const core::likec::FunctionDefinition *getFunctionDefinition(const core::likec::FunctionDeclaration *declaration) const {
-        assert(declaration != nullptr);
-        return nc::find(functionDeclaration2definition_, declaration);
-    }
-
-    /**
      * Replaces the text of all identifiers referring to the given declaration
      * with the given name.
      *
@@ -147,25 +100,6 @@ public:
      * \return Text in the given range.
      */
     QString getText(const Range<int> &range) const;
-
-    /**
-     * For a node, computes statement, term, and instruction, from which
-     * this node has originated.
-     *
-     * \param[in]  node         Valid pointer to a tree node.
-     * \param[out] statement    Original statement.
-     * \param[out] term         Original term.
-     * \param[out] instruction  Original instruction.
-     */
-    static void getOrigin(const core::likec::TreeNode *node, const core::ir::Statement *&statement,
-                          const core::ir::Term *&term, const core::arch::Instruction *&instruction);
-
-    /**
-     * \param node Valid pointer to a tree node.
-     *
-     * \return Declaration of whatever identifier this node is. Can be nullptr.
-     */
-    static const core::likec::Declaration *getDeclarationOfIdentifier(const core::likec::TreeNode *node);
 
 private Q_SLOTS:
     void onContentsChange(int position, int charsRemoved, int charsAdded);
